sssd.conf - original check added for checking if the network is up. The operstate check is carried out first followed by the ping check
setup
Added in a grub configuration to stop the boot menu from appearing.

set-network
Now sets the time to the pdc before joining the domain.

This fixes a problem where the time has been set in the bios to British
Summer Time.
On bootup the client joins the domain and then ntp sets the time one hour out
 which causes problems on the intial reboot after joining the domain with
kerberos authentication failing.